remove the package 'ml/strategy', rename 'ml/interface' to 'ml/strategies', update the class Worker
