--- src/syngen/ml/interface/__init__.py
@@ -1,2 +1,0 @@
-from syngen.ml.interface.interface import TrainInterface
-from syngen.ml.interface.interface import InferInterface
